animate and functions moved back to main
